Kotlin: DB classes completed

+++ app/src/main/kotlin/ru/vinyarsky/englishmedia/models/data/db/EpisodeEntity.kt
@@ -4,6 +4,21 @@ import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
+/**
+ * Episode database entity
+ *
+ * @property code Unique identifier
+ * @property podcastCode Unique identifier of a podcast, this episode belongs to
+ * @property episodeGuid Episode identifier in an rss feed
+ * @property title Title
+ * @property description Description
+ * @property pageUrl Webpage URL
+ * @property contentUrl Audio file URL
+ * @property duration Duration (in seconds)
+ * @property pubDate Date of publishing
+ * @property status Status of listening
+ * @property currentPosition Current position (seconds from the beginning)
+ */
 @Entity(tableName = "Episodes")
 data class EpisodeEntity(
         @PrimaryKey
@@ -17,6 +32,6 @@ data class EpisodeEntity(
         @ColumnInfo(name = "page_url") var pageUrl: String?,
         @ColumnInfo(name = "content_url") var contentUrl: String,
         @ColumnInfo(name = "duration") var duration: Int,
-        @ColumnInfo(name = "pub_date") var pubDate: Int?,
+        @ColumnInfo(name = "pub_date") var pubDate: Long?,
         @ColumnInfo(name = "status") var status: String,
         @ColumnInfo(name = "current_position") var currentPosition: Int = 0)
